% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_facet.R
\name{facet_date_range}
\alias{facet_date_range}
\title{Specify a date range facet}
\usage{
facet_date_range(query, field = c("pubdate", "indexdate"), start, end,
  gap)
}
\description{
Specify a date range facet
}
