% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_guid}
\alias{filter_guid}
\title{Filter guid}
\usage{
filter_guid(query, value)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}} or \code{\link[=query_facet]{query_facet()}} object}

\item{value}{guid to filter}
}
\description{
Filter guid
}
