% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_text}
\alias{filter_text}
\title{Filter based on text found in title, description, or text, including
translations}
\usage{
filter_text(query, text)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}} or \code{\link[=query_facet]{query_facet()}} object}

\item{text}{text to filter}
}
\description{
Filter based on text found in title, description, or text, including
translations
}
