% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_tonality}
\alias{filter_tonality}
\title{Filter tonality}
\usage{
filter_tonality(query, value = NULL, from = NULL, to = NULL)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}} or \code{\link[=query_facet]{query_facet()}} object}

\item{value}{exact tonality value to filter on (if specified, from and to are ignored)}

\item{from}{lower range to filter on}

\item{to}{upper range to filter on}
}
\description{
Filter tonality
}
