% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{finder_connect}
\alias{finder_connect}
\title{Set connection details to Finder engine}
\usage{
finder_connect(host = "127.0.0.1", port = NULL)
}
\arguments{
\item{host}{The base host, defaults to 127.0.0.1}

\item{port}{The port to connect to (optional)}
}
\description{
Set connection details to Finder engine
}
\examples{
\dontrun{
con <- finder_connect(host = "localhost")
}
}
