% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_date_filter}
\alias{fq_date_filter}
\title{Filter on a generic date field}
\usage{
fq_date_filter(query, field, from = NULL, to = NULL, on = NULL)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{field}{the name of the date field to filter on}

\item{from}{start date}

\item{to}{end date}

\item{on}{exact date (if specified, from and to are ignored)}
}
\description{
Filter on a generic date field
}
