% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_facet.R
\name{fq_facet_date_range}
\alias{fq_facet_date_range}
\title{Specify a date range facet}
\usage{
fq_facet_date_range(query, field = c("pubdate", "indexdate"), start, end, gap)
}
\arguments{
\item{query}{a \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{field}{field name (one of "pubdate" or "indexdate")}

\item{start}{start date}

\item{end}{end date}

\item{gap}{gap object from \code{\link[=fq_range_gap]{fq_range_gap()}}}
}
\description{
Specify a date range facet
}
