% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_filter_duplicate}
\alias{fq_filter_duplicate}
\title{Filter duplicate}
\usage{
fq_filter_duplicate(query, bool = FALSE)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{bool}{TRUE or FALSE}
}
\description{
Filter duplicate
}
