% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_filter_entityid}
\alias{fq_filter_entityid}
\title{Filter entityid}
\usage{
fq_filter_entityid(query, terms)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{terms}{A vector of search terms. If preceded with a "+", (e.g. "+de"), then this term is required to be in the result. If preceded with "-", (e.g. "-en"), then this term is prohibited from appearing in the result.}
}
\description{
Filter entityid
}
