% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_filter_guid}
\alias{fq_filter_guid}
\title{Filter guid}
\usage{
fq_filter_guid(query, value)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{value}{guid to filter}
}
\description{
Filter guid
}
