% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_filter_indexdate}
\alias{fq_filter_indexdate}
\title{Filter indexdate}
\usage{
fq_filter_indexdate(query, from = NULL, to = NULL, on = NULL)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{from}{start date}

\item{to}{end date}

\item{on}{exact date (if specified, from and to are ignored)}
}
\description{
Filter indexdate
}
