% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_filter_text}
\alias{fq_filter_text}
\title{Filter based on text found in title, description, or text, including
translations}
\usage{
fq_filter_text(query, text)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{text}{text to filter}
}
\description{
Filter based on text found in title, description, or text, including
translations
}
