% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fq_get_query}
\alias{fq_get_query}
\title{Get the query string for a query}
\usage{
fq_get_query(query)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}}, \code{\link[=fq_query_facet]{fq_query_facet()}}, or \code{\link[=fq_query_str]{fq_query_str()}} object}
}
\description{
Get the query string for a query
}
