% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{fq_n_docs}
\alias{fq_n_docs}
\title{Get how many documents were returned}
\usage{
fq_n_docs(x)
}
\arguments{
\item{x}{Documents object in list format}
}
\description{
Get how many documents were returned
}
