% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_q}
\alias{fq_q}
\title{Explicitly specify the "q=..." part of a query}
\usage{
fq_q(query, q)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{q}{text to filter}
}
\description{
Explicitly specify the "q=..." part of a query
}
