% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_facet.R
\name{fq_query_facet}
\alias{fq_query_facet}
\title{Initialize a document fetch query}
\usage{
fq_query_facet(con)
}
\arguments{
\item{con}{A finder connection object from \code{\link[=finder_connect]{finder_connect()}}.}
}
\description{
Initialize a document fetch query
}
