% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fetch.R
\name{fq_query_fetch}
\alias{fq_query_fetch}
\title{Initialize a document fetch query}
\usage{
fq_query_fetch(
  con,
  path = NULL,
  max = 10,
  size = 10000,
  format = c("list", "xml", "file")
)
}
\arguments{
\item{con}{A finder connection object from \code{\link[=finder_connect]{finder_connect()}}.}

\item{path}{An optional directory in which to place downloaded results
(required when \code{format="file"}).}

\item{max}{The maximum number of documents to fetch. Set to <0 to fetch
all. When fetching all it is a good idea to first fetch with only a few
documents and use \code{\link[=n_docs]{n_docs()}} to get a sense of how many will be pulled.}

\item{size}{The number of documents to fetch in each batch (max is 10000).}

\item{format}{One of "list", "xml", "file". In the case of "file",
The path to the file will be returned. In all other cases, an object
with the fetched data will be returned in the format specified. List
results are converted using \code{\link[=fq_xml_to_list]{fq_xml_to_list()}}.}
}
\description{
Initialize a document fetch query
}
