% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_str.R
\name{fq_query_str}
\alias{fq_query_str}
\title{Initialize a query with a specified query string}
\usage{
fq_query_str(
  con,
  str,
  path = NULL,
  format = c("list", "xml", "file"),
  type = c("unknown", "fetch", "facet"),
  max = 10,
  size = 10000
)
}
\arguments{
\item{con}{A finder connection object from \code{\link[=finder_connect]{finder_connect()}}.}

\item{str}{A SOLR query string, e.g. "op=search&q=\emph{:}".}

\item{path}{An optional directory in which to place downloaded results
(required when \code{format="file"}).}

\item{format}{One of "list", "xml", "file". In the case of "file",
The path to the file will be returned. In all other cases, an object
with the fetched data will be returned in the format specified. List
results are converted using \code{\link[=fq_xml_to_list]{fq_xml_to_list()}}.}

\item{type}{the type of query defined in the string - one of "unknown",
"fetch", or "agg"}

\item{max}{The maximum number of documents to fetch (used when
type=="fetch"). Set to <0 to fetch all. When fetching all it is a
good idea to first fetch with only a few documents and use \code{\link[=fq_n_docs]{fq_n_docs()}}
to get a sense of how many will be pulled.}

\item{size}{The number of documents to fetch in each batch (max is
10000 - used when type=="fetch").}
}
\description{
Initialize a query with a specified query string
}
