% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_facet.R
\name{fq_range_gap}
\alias{fq_range_gap}
\title{Specify a 'gap' parameter for facet_range queries}
\usage{
fq_range_gap(
  num = 1,
  units = c("DAY", "MINUTE", "HOUR", "WEEK", "MONTH", "YEAR")
)
}
\arguments{
\item{num}{the number of units to use for a date gap}

\item{units}{the units to use for a date gap, one of "DAY", "MINUTE", "HOUR", "WEEK", "MONTH", "YEAR"}
}
\description{
Specify a 'gap' parameter for facet_range queries
}
