% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_select.R
\name{fq_select_fields}
\alias{fq_select_fields}
\title{Specify fields to return in the results}
\usage{
fq_select_fields(query, fields = NULL)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} object}

\item{fields}{a vector of field names (see \code{\link[=fq_selectable_fields]{fq_selectable_fields()}})}
}
\description{
Specify fields to return in the results
}
