% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_select.R
\name{fq_sort_by}
\alias{fq_sort_by}
\title{Specify field to sort on}
\usage{
fq_sort_by(query, field, asc = TRUE)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} object}

\item{field}{the field name to sort by (see \code{\link[=fq_queryable_fields]{fq_queryable_fields()}})}

\item{asc}{if TRUE, will sort ascending, if FALSE, descending}
}
\description{
Specify field to sort on
}
