% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{fq_term_filter}
\alias{fq_term_filter}
\title{Filter on a generic field}
\usage{
fq_term_filter(query, field, terms)
}
\arguments{
\item{query}{a \code{\link[=fq_query_fetch]{fq_query_fetch()}} or \code{\link[=fq_query_facet]{fq_query_facet()}} object}

\item{field}{the name of the field to filter on}

\item{terms}{exact value(s) to filter theid field on}
}
\description{
Filter on a generic field
}
