% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_values.R
\name{fq_valid_languages}
\alias{fq_valid_languages}
\title{Get a vector of all unique "language" values}
\usage{
fq_valid_languages(con)
}
\arguments{
\item{con}{A finder connection object from \code{\link[=finder_connect]{finder_connect()}}.}
}
\description{
Get a vector of all unique "language" values
}
