% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{fq_xml_to_list}
\alias{fq_xml_to_list}
\title{Convert Finder xml result into a more convenient R list format}
\usage{
fq_xml_to_list(x)
}
\arguments{
\item{x}{an object of class "xml_document" obtained from \code{\link[=fq_query_fetch]{fq_query_fetch()}}}
}
\description{
Convert Finder xml result into a more convenient R list format
}
