% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_query}
\alias{get_query}
\title{Get the query string for a query}
\usage{
get_query(query)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}}, \code{\link[=query_facet]{query_facet()}}, or \code{\link[=query_str]{query_str()}} object}
}
\description{
Get the query string for a query
}
