% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert documents list to data frame}
\usage{
list_to_df(x)
}
\arguments{
\item{x}{an object of class "xml_document" or "finder_docs" obtained from \code{\link[=query_fetch]{query_fetch()}}.}
}
\description{
Convert documents list to data frame
}
\details{
In general, a lot of information can be lost when transforming to a data frame, such as attributes or entries that have more than one element. Columns "title" and "description" often have two elements with the second being the English translation. Because of this, columns "title", "title_en", "description", and "description_en" are created to preserve this. Columns "entity", "category", "fullgeo" usually have more than one entry for each document and as such are added to the data frame as "list columns".
}
