% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fetch.R
\name{query_fetch}
\alias{query_fetch}
\title{Initialize a document fetch query}
\usage{
query_fetch(con, str, path = NULL, max = 0, size = 10000,
  format = c("list", "json", "xml", "file"))
}
\arguments{
\item{con}{A finder connection object from [finder_connect()].}

\item{str}{A SOLR query string, e.g. "op=search&q=*:*&rows=0".}

\item{path}{An optional directory in which to place downloaded results
(required when `format="file"`).}

\item{max}{The maximum number of documents to fetch.}

\item{size}{The number of documents to fetch in each batch (max is 10000).}

\item{format}{One of "list", "json", "xml", "file". In the case of "file",
The path to the file will be returned. In all other cases, an object
with the fetched data will be returned in the format specified.}
}
\description{
Initialize a document fetch query
}
