% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_facet.R
\name{range_gap}
\alias{range_gap}
\title{Specify a 'gap' parameter for facet_range queries}
\usage{
range_gap(num = 1, units = c("DAY", "MINUTE", "HOUR", "WEEK", "MONTH", "YEAR"))
}
\description{
Specify a 'gap' parameter for facet_range queries
}
