% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_select.R
\name{select_fields}
\alias{select_fields}
\title{Specify fields to return in the results}
\usage{
select_fields(query, fields = NULL)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}} object}

\item{fields}{a vector of field names (see \code{\link[=selectable_fields]{selectable_fields()}})}
}
\description{
Specify fields to return in the results
}
