% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_select.R
\name{sort_by}
\alias{sort_by}
\title{Specify field to sort on}
\usage{
sort_by(query, field, asc = TRUE)
}
\arguments{
\item{query}{a \code{\link[=query_fetch]{query_fetch()}} object}

\item{field}{the field name to sort by (see \code{\link[=queryable_fields]{queryable_fields()}})}

\item{asc}{if TRUE, will sort ascending, if FALSE, descending}
}
\description{
Specify field to sort on
}
