% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_values.R
\name{valid_categories}
\alias{valid_categories}
\title{Get a vector of all unique "category" values}
\usage{
valid_categories(con)
}
\arguments{
\item{con}{A finder connection object from \code{\link[=finder_connect]{finder_connect()}}.}
}
\description{
Get a vector of all unique "category" values
}
