% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_values.R
\name{valid_duplicate}
\alias{valid_duplicate}
\title{Get a vector of all unique "duplicate" values}
\usage{
valid_duplicate(con)
}
\arguments{
\item{con}{A finder connection object from \code{\link[=finder_connect]{finder_connect()}}.}
}
\description{
Get a vector of all unique "duplicate" values
}
