% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_values.R
\name{valid_quotecategory}
\alias{valid_quotecategory}
\title{Get a vector of all unique "quotecategory" values}
\usage{
valid_quotecategory(con)
}
\arguments{
\item{con}{A finder connection object from \code{\link[=finder_connect]{finder_connect()}}.}
}
\description{
Get a vector of all unique "quotecategory" values
}
