% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{write_docs_csv}
\alias{write_docs_csv}
\title{Write xml documents to a csv file}
\usage{
write_docs_csv(x, collapse = ";", ...)
}
\arguments{
\item{x}{an object of class "xml_document" or "finder_docs" obtained from \code{\link[=query_fetch]{query_fetch()}}.}

\item{collapse}{The separator character to use to collapse list column fields.}

\item{\ldots}{Parameters sent to \code{\link[readr:write_delim]{readr::write_csv()}}.}
}
\description{
Write xml documents to a csv file
}
\details{
List columns of the data frame will be collapsed. Currently these columns are "entity", "category", "fullgeo"
}
