% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{xml_to_list}
\alias{xml_to_list}
\title{Convert Finder xml result into a more convenient R list format}
\usage{
xml_to_list(x, fields = NULL)
}
\description{
Convert Finder xml result into a more convenient R list format
}
